+++ service/src/main/java/com/wilma/service/forum/CategoryService.java
@@ -2,6 +2,7 @@ package com.wilma.service.forum;
 
 import com.wilma.entity.forum.ForumCategory;
 import com.wilma.repository.CategoryRepository;
+import com.wilma.service.CrudOpsImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
